Move Help and About to src/components/help-etc
