+++ src/components/help-etc/About.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import NavRoutes from './NavRoutes';
-import ClimateExplorer from "../help-etc/ClimateExplorer";
+import NavRoutes from '../navigation/NavRoutes/NavRoutes';
+import ClimateExplorer from "./ClimateExplorer";
 
 const navSpec = {
   basePath: '/about',
